requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-WTF==1.2.2
flask-babel==4.0.0
python-dotenv==1.0.1

celery==5.4.0
redis==5.2.1
APScheduler==3.11.0

pandas==2.2.3
numpy==2.2.1
openpyxl==3.1.5
XlsxWriter==3.2.0
python-docx==1.1.2

PyYAML==6.0.2
gunicorn==23.0.0
